+++ CHANGELOG.md
@@ -55,6 +55,18 @@ These classes will run as threads in the next releases:
 - all plugins in bridge mode;
 - each plugin in childbridge mode;
 
+## [3.7.7] - Dev branch
+
+### Changed
+
+- [package]: Update jest config to v.2.0.0, prettier config to v.2.0.0 and eslint config to v.2.0.0.
+
+### Fixed
+
+- [frontend]: Fix add plugin after upload.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
 ## [3.7.6] - 2026-05-01
 
 ### Development changes